+++ engine/blackjack.go
@@ -14,12 +14,12 @@ type Blackjack struct {
 	players    []*Player
 	discard    *game.Deck
 	maxDiscard int
-	minBet     int
-	maxBet     int
+	minBet     float64
+	maxBet     float64
 }
 
 // NewBlackjack will create a new game engine.
-func NewBlackjack(numDecks int, maxDiscard int, minBet int, maxBet int, dealer *Player, players ...*Player) *Blackjack {
+func NewBlackjack(numDecks int, maxDiscard int, minBet float64, maxBet float64, dealer *Player, players ...*Player) *Blackjack {
 	shuffler := game.NewShuffler()
 
 	deck := game.NewDeck()
@@ -89,12 +89,12 @@ func (b *Blackjack) DisplayStats() {
 }
 
 func (b *Blackjack) placeBet(player *Player) {
-	fmt.Printf("%s has $%d. Min bet is $%d and max bet is $%d.\n", player.Name, player.Money, b.minBet, b.maxBet)
+	fmt.Printf("%s has $%.2f. Min bet is $%.2f and max bet is $%.2f.\n", player.Name, player.Money, b.minBet, b.maxBet)
 	player.Bet = player.AI.PlaceBet(b.minBet, b.maxBet, player.Money)
 	if player.Bet >= b.minBet && player.Bet <= b.maxBet {
-		fmt.Printf("%s placed a bet of $%d.\n", player.Name, player.Bet)
+		fmt.Printf("%s placed a bet of $%.2f.\n", player.Name, player.Bet)
 	} else {
-		fmt.Printf("%s tried to place an invalid bet of $%d. Will use minimum bet of $%d.\n", player.Name, player.Bet, b.minBet)
+		fmt.Printf("%s tried to place an invalid bet of $%.2f. Will use minimum bet of $%.2f.\n", player.Name, player.Bet, b.minBet)
 		player.Bet = b.minBet
 	}
 }
@@ -139,7 +139,7 @@ func (b *Blackjack) playerTurn(player *Player) bool {
 			}
 			card := b.dealCard(player.Hand, false)
 			player.Bet *= 2
-			fmt.Printf("%s doubled their bet to $%d and was dealt: %v\n", player.Name, player.Bet, card)
+			fmt.Printf("%s doubled their bet to $%.2f and was dealt: %v\n", player.Name, player.Bet, card)
 			b.displayHand(player.Name, player.Hand)
 		case game.ActionStats:
 			b.displayPlayerStats(player)
@@ -182,7 +182,7 @@ func (b *Blackjack) determineWinners() {
 		if p.Hand.IsNatural() {
 			fmt.Printf("%s has a natural blackjack!\n", p.Name)
 			p.Win++
-			p.Money += int(float32(p.Bet) * 1.5) // TODO: figure out what to do about truncation? Make it a float?
+			p.Money += p.Bet * 1.5
 		} else if playerTotal > 21 {
 			fmt.Printf("%s busted with a total of %d.\n", p.Name, playerTotal)
 			p.Loss++
@@ -238,7 +238,7 @@ func (b *Blackjack) displayHand(name string, hand *game.Hand) {
 func (b *Blackjack) displayPlayerStats(player *Player) {
 	total := player.Win + player.Loss + player.Tie
 	fmt.Printf("%s (%T)\n", player.Name, player.AI)
-	fmt.Printf("  Win: %d (%%%.1f) | Loss: %d (%%%.1f) | Tie: %d (%%%.1f) | $%d\n", player.Win, percent(player.Win, total), player.Loss, percent(player.Loss, total), player.Tie, percent(player.Tie, total), player.Money)
+	fmt.Printf("  Win: %d (%%%.1f) | Loss: %d (%%%.1f) | Tie: %d (%%%.1f) | $%.2f\n", player.Win, percent(player.Win, total), player.Loss, percent(player.Loss, total), player.Tie, percent(player.Tie, total), player.Money)
 }
 
 // percent will calculate a percentage from the given values.
